Included update end-point

+++ src/main/java/com/leandropitta/cost_management/dto/response/CostResponseDto.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Builder
@@ -14,7 +13,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CostResponseDto {
-    private Integer id;
+    private Long id;
     private LocalDateTime date;
     private String buy;
     private BigDecimal cost;
